+++ src/js/todo.js
@@ -1,88 +1,17 @@
 import { formatDistance } from 'date-fns';
 import auth from './auth';
 import store from './store';
+import DragAndDrop from './dragdrop';
 
 const errorContainer = document.getElementById('email-error');
 
 const todoStore = store();
+const todoDragAndDrop = DragAndDrop();
 
 const toggleCompleted = (todoId, refreshTodo) => {
   const { userId } = auth.getSession();
   todoStore.toggleComplete(userId, todoId);
   refreshTodo();
-};
-
-let draggedElement = null;
-
-const handleDragStart = (event) => {
-  draggedElement = event.target;
-  event.target.classList.add('bg-indigo-300');
-};
-
-const handleDragOver = (event) => {
-  event.preventDefault();
-  const hoveredElement = event.target;
-
-  if (hoveredElement !== draggedElement) {
-    const hoveredRect = hoveredElement.getBoundingClientRect();
-    const draggedRect = draggedElement.getBoundingClientRect();
-
-    const distance = hoveredRect.top - draggedRect.top;
-
-    if (distance > 0) {
-      hoveredElement.parentNode.insertBefore(draggedElement, hoveredElement.nextSibling);
-    } else {
-      hoveredElement.parentNode.insertBefore(draggedElement, hoveredElement);
-    }
-  }
-
-  const draggedElementRect = event.target.getBoundingClientRect();
-  const topOffset = draggedElementRect.top;
-  const bottomOffset = window.innerHeight - draggedElementRect.bottom;
-
-  const scrollThreshold = 150;
-
-  if (topOffset < scrollThreshold) {
-    window.scrollTo({
-      top: window.scrollY - (scrollThreshold - topOffset),
-      behavior: 'smooth',
-    });
-  }
-
-  if (bottomOffset < scrollThreshold) {
-    window.scrollTo({
-      top: window.scrollY + (scrollThreshold - bottomOffset),
-      behavior: 'smooth',
-    });
-  }
-};
-
-const handleDragEnd = (event) => {
-  event.target.classList.remove('bg-indigo-300');
-};
-
-const handleDrop = (event) => {
-  event.preventDefault();
-  const todoId = event.dataTransfer.getData('text/plain');
-  const draggedTodoItem = document.getElementById(todoId);
-  const todoList = document.getElementById('todoList');
-
-  if (draggedTodoItem) {
-    const closestTodoItem = event.target.closest('.todo-item');
-
-    if (closestTodoItem) {
-      if (
-        event.clientY <
-        closestTodoItem.getBoundingClientRect().top + closestTodoItem.offsetHeight / 2
-      ) {
-        todoList.insertBefore(draggedTodoItem, closestTodoItem);
-      } else {
-        todoList.insertBefore(draggedTodoItem, closestTodoItem.nextSibling);
-      }
-    } else {
-      todoList.appendChild(draggedTodoItem);
-    }
-  }
 };
 
 const renderTodo = (todo) => {
@@ -130,9 +59,6 @@ const renderTodo = (todo) => {
     </div>
   `;
 
-  li.addEventListener('dragstart', handleDragStart);
-  li.addEventListener('dragend', handleDragEnd);
-
   return li;
 };
 
@@ -152,8 +78,10 @@ const renderTodos = () => {
     });
   });
 
-  todoList.addEventListener('dragover', handleDragOver);
-  todoList.addEventListener('drop', handleDrop);
+  todoList.addEventListener('dragstart', todoDragAndDrop.handleDragStart);
+  todoList.addEventListener('dragover', todoDragAndDrop.handleDragOver);
+  todoList.addEventListener('dragend', todoDragAndDrop.handleDragEnd);
+  todoList.addEventListener('drop', todoDragAndDrop.handleDrop);
 
   document.querySelector('#todo-count').textContent =
     todos.length > 1 ? `${todos.length} items` : `${todos.length} item`;
